# frontend/src/api-client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "YAESandBox API (Public)",
    "description": "供前端和外部使用的公开 API 文档。",
    "version": "v1"
  },
  "paths": {
...
    "/api/entities": {
      "get": {
        "tags": [
          "Entities"
        ],
        "summary": "获取指定 Block 当前可交互 WorldState 中的所有非销毁实体摘要信息。",
        "parameters": [
          {
            "name": "blockId",
            "in": "query",
            "description": "要查询的目标 Block 的 ID。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回实体摘要列表。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntitySummaryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntitySummaryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntitySummaryDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "缺少必需的 'blockId' 查询参数。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "未找到具有指定 ID 的 Block 或 Block 无法访问。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/entities/{entityType}/{entityId}": {
      "get": {
        "tags": [
          "Entities"
        ],
        "summary": "获取指定 Block 当前可交互 WorldState 中的单个非销毁实体的详细信息。",
        "parameters": [
          {
            "name": "entityType",
            "in": "path",
            "description": "要查询的实体的类型 (Item, Character, Place)。",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityType"
            }
          },
          {
            "name": "entityId",
            "in": "path",
            "description": "要查询的实体的 ID。",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "blockId",
            "in": "query",
            "description": "目标 Block 的 ID。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回实体详细信息。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "缺少必需的 'blockId' 查询参数。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "未在指定 Block 中找到实体，或 Block 未找到。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/blocks/{blockId}/GameState": {
      "get": {
        "tags": [
          "GameState"
        ],
        "summary": "获取指定 Block 的当前 GameState。",
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "description": "目标 Block 的 ID。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回 GameState。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateDto"
                }
              }
            }
          },
          "404": {
            "description": "未找到具有指定 ID 的 Block。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "GameState"
        ],
        "summary": "修改指定 Block 的 GameState。使用 PATCH 方法进行部分更新。",
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "description": "目标 Block 的 ID。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "包含要更新的 GameState 键值对的请求体。",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameStateRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameStateRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameStateRequestDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "GameState 更新成功。"
          },
          "400": {
            "description": "请求体无效。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "未找到具有指定 ID 的 Block。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "更新时发生内部服务器错误。"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntityDetailDto": {
        "type": "object",
        "properties": {
          "entityId": {
            "type": "string",
            "description": "实体的唯一 ID。",
            "nullable": true
          },
          "entityType": {
            "$ref": "#/components/schemas/EntityType"
          },
          "isDestroyed": {
            "type": "boolean",
            "description": "指示实体是否已被标记为销毁。\r\n注意：查询 API 通常只返回未销毁的实体。"
          },
          "name": {
            "type": "string",
            "description": "实体的名称 (通常来自 'name' 属性，如果不存在则可能回退到 EntityId)。",
            "nullable": true
          },
          "attributes": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "description": "包含实体所有属性（包括核心属性如 IsDestroyed 和动态属性）的字典。\r\n值的类型可能是 string, int, bool, double, List[object?], Dictionary-[string, object?], TypedID 等。",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "用于 API 响应，表示实体的详细信息，包含所有属性。\r\n继承自 YAESandBox.API.DTOs.EntitySummaryDto。"
      },
      "EntitySummaryDto": {
        "type": "object",
        "properties": {
          "entityId": {
            "type": "string",
            "description": "实体的唯一 ID。",
            "nullable": true
          },
          "entityType": {
            "$ref": "#/components/schemas/EntityType"
          },
          "isDestroyed": {
            "type": "boolean",
            "description": "指示实体是否已被标记为销毁。\r\n注意：查询 API 通常只返回未销毁的实体。"
          },
          "name": {
            "type": "string",
            "description": "实体的名称 (通常来自 'name' 属性，如果不存在则可能回退到 EntityId)。",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "用于 API 响应，表示实体的基本摘要信息。"
      },
      "EntityType": {
        "enum": [
          "Item",
          "Character",
          "Place"
        ],
        "type": "string"
      },
      "GameStateDto": {
        "type": "object",
        "properties": {
          "settings": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "description": "包含 GameState 所有设置的字典。\r\n键是设置的名称 (string)，值是设置的值 (object?)。\r\n值的实际类型取决于具体的游戏状态设置。",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "用于 API 响应，表示一个 Block 的 GameState。"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "UpdateGameStateRequestDto": {
        "required": [
          "settingsToUpdate"
        ],
        "type": "object",
        "properties": {
          "settingsToUpdate": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "description": "一个字典，包含要更新或添加的 GameState 设置。\r\n键是要修改的设置名称，值是新的设置值。\r\n如果值为 null，通常表示移除该设置或将其设置为空。"
          }
        },
        "additionalProperties": false,
        "description": "用于 API 请求，表示要更新的 GameState 设置。"
      },
...
    }
  }
}